Added licensing information

+++ firpm_ld/src/cheby.cpp
@@ -1,9 +1,19 @@
-/*
- * cheby.cpp
- *
- *  Created on: Sep 26, 2013
- *      Author: silviu
- */
+//    firpm_ld
+//    Copyright (C) 2015  S. Filip
+//
+//    This program is free software: you can redistribute it and/or modify
+//    it under the terms of the GNU General Public License as published by
+//    the Free Software Foundation, either version 3 of the License, or
+//    (at your option) any later version.
+
+//    This program is distributed in the hope that it will be useful,
+//    but WITHOUT ANY WARRANTY; without even the implied warranty of
+//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//    GNU General Public License for more details.
+
+//    You should have received a copy of the GNU General Public License
+//    along with this program.  If not, see <http://www.gnu.org/licenses/>
+
 
 #include "firpm/cheby.h"
 
@@ -103,7 +113,7 @@ void generateChebyshevPoints(std::vector<long double>& x, std::size_t n)
     x.reserve(n + 1u);
     if(n > 0u)
     {
-        for(int k = n; k >= -n; k -= 2)
+        for(int k = n; k >= -(int)n; k -= 2)
             x.push_back(sinl(M_PI * k / (n * 2)));
     }
     else
